+++ src/ZKEACMS.Product/Service/ProductService.cs
@@ -73,7 +73,7 @@ namespace ZKEACMS.Product.Service
                         item.ProductImages.Each(m =>
                         {
                             m.ProductId = item.ID;
-                            if (m.ActionType == ActionType.Create)
+                            if (m.ActionType.HasFlag(ActionType.Create))
                             {
                                 _productImageService.Add(m);
                             }
@@ -86,27 +86,22 @@ namespace ZKEACMS.Product.Service
         }
         private void SaveImages(ProductImage item)
         {
-            switch (item.ActionType)
+            if (item.ActionType.HasFlag(ActionType.Create))
             {
-                case ActionType.Create:
-                    {
-                        _productImageService.Add(item);
-                        break;
-                    }
-                case ActionType.Update:
-                    {
-                        _productImageService.Update(item);
-                        break;
-                    }
-                case ActionType.Delete:
-                    {
-                        if (item.ID > 0)
-                        {
-                            _productImageService.Remove(item);
-                        }
-                        break;
-                    }
+                _productImageService.Add(item);
             }
+            else if (item.ActionType.HasFlag(ActionType.Update))
+            {
+                _productImageService.Update(item);
+            }
+            else if (item.ActionType.HasFlag(ActionType.Delete))
+            {
+                if (item.ID > 0)
+                {
+                    _productImageService.Remove(item);
+                }
+            }
+
         }
         public override ServiceResult<ProductEntity> Update(ProductEntity item)
         {
@@ -196,7 +191,7 @@ namespace ZKEACMS.Product.Service
         }
         public ProductEntity GetByUrl(string url)
         {
-            ProductEntity product= Get(m => m.Url == url).FirstOrDefault();
+            ProductEntity product = Get(m => m.Url == url).FirstOrDefault();
             if (product != null)
             {
                 product.ProductTags = _productCategoryTagService.Get(m => m.ProductCategoryId == product.ProductCategoryID);
